server/requirements.txt
mido==1.2.10
python-rtmidi==1.4.9
redis==3.5.3
